+++ src/governor/ddl.rs
@@ -1,5 +1,7 @@
 use crate::cas;
 use crate::governor::backup::handle_backup;
+use crate::governor::batch::get_next_batch;
+use crate::governor::batch::LastStepInfo;
 use crate::governor::plan::stage::Plan;
 use crate::governor::plan::stage::*;
 use crate::has_states;
@@ -21,7 +23,8 @@ use crate::traft::RaftTerm;
 use crate::traft::Result;
 use crate::vshard;
 use crate::warn_or_panic;
-use futures::future::try_join_all;
+use futures::stream::FuturesUnordered;
+use futures::stream::StreamExt;
 use std::collections::HashMap;
 use std::time::Duration;
 use tarantool::space::SpaceId;
@@ -36,16 +39,23 @@ use tarantool::space::SpaceId;
 /// - TRUNCATE TABLE operation (implemented as DDL in tarantool)
 /// - BACKUP operation
 pub fn handle_pending_ddl<'i>(
+    last_step_info: &mut LastStepInfo,
     topology_ref: &TopologyCacheRef,
     tables: &HashMap<SpaceId, &'i TableDef>,
     pending_ddl: &'i Option<(Ddl, u64)>,
+    schema_version: u64,
     term: RaftTerm,
     applied: RaftIndex,
     sync_timeout: Duration,
+    batch_size: usize,
 ) -> Result<Option<Plan<'i>>> {
     let Some((ref ddl, pending_schema_version)) = *pending_ddl else {
         return Ok(None);
     };
+
+    // Clear previous results if schema version changed, because it means that
+    // we're applying a different operation
+    last_step_info.update_schema_info(schema_version, pending_schema_version);
 
     if let Ddl::Backup { timestamp } = ddl {
         return handle_backup(
@@ -60,17 +70,36 @@ pub fn handle_pending_ddl<'i>(
 
     if let Ddl::TruncateTable { id, .. } = ddl {
         return handle_truncate_table(
+            last_step_info,
             topology_ref,
             tables,
+            schema_version,
             pending_schema_version,
             *id,
             term,
             applied,
             sync_timeout,
+            batch_size,
         );
     }
 
-    let targets = rpc::replicasets_masters(topology_ref);
+    // We must check if step kind is different from the one we tried on
+    // previous iteration, so that we know not to use irrelevant results
+    type Action<'a> = ApplySchemaChange<'a>;
+    let step_kind = Action::KIND;
+    last_step_info.update_step_kind(step_kind);
+
+    let targets_total = rpc::replicasets_masters(topology_ref);
+    debug_assert!(!targets_total.is_empty());
+    tlog!(Info, "targets_total: {targets_total:?}");
+
+    let res = get_next_batch(&targets_total, last_step_info, batch_size);
+    let targets_batch = match res {
+        Ok(v) => v,
+        Err(next_try) => {
+            return Ok(Some(SleepDueToBackoff::new(next_try, step_kind).into()));
+        }
+    };
 
     let rpc = rpc::ddl_apply::Request {
         term,
@@ -79,11 +108,13 @@ pub fn handle_pending_ddl<'i>(
     };
 
     Ok(Some(
-        ApplySchemaChange {
+        Action {
             ddl,
             pending_schema_version,
+            schema_version,
             rpc,
-            targets,
+            targets_total,
+            targets_batch,
         }
         .into(),
     ))
@@ -104,13 +135,16 @@ pub fn handle_pending_ddl<'i>(
 ///
 /// This function is called from [`handle_pending_ddl`]
 fn handle_truncate_table<'i>(
+    last_step_info: &mut LastStepInfo,
     topology_ref: &TopologyCacheRef,
     tables: &HashMap<SpaceId, &'i TableDef>,
+    schema_version: u64,
     pending_schema_version: u64,
     table_id: SpaceId,
     term: RaftTerm,
     applied: RaftIndex,
     sync_timeout: Duration,
+    batch_size: usize,
 ) -> Result<Option<Plan<'i>>> {
     let table_def = tables.get(&table_id).expect("failed to get table_def");
     let tier = table_def.distribution.in_tier().cloned();
@@ -131,8 +165,14 @@ fn handle_truncate_table<'i>(
         ));
     };
 
+    // We must check if step kind is different from the one we tried on
+    // previous iteration, so that we know not to use irrelevant results
+    type Action = ApplyTruncateTable;
+    let step_kind = Action::KIND;
+    last_step_info.update_step_kind(step_kind);
+
     let mut tier_masters_count = 0;
-    let mut other_tiers_masters = vec![];
+    let mut other_tiers_masters_total = vec![];
 
     for replicaset in topology_ref.all_replicasets() {
         let Ok(master) = topology_ref.instance_by_name(&replicaset.current_master_name) else {
@@ -145,7 +185,7 @@ fn handle_truncate_table<'i>(
             if replicaset.tier == tier {
                 tier_masters_count += 1;
             } else {
-                other_tiers_masters.push(replicaset.current_master_name.clone());
+                other_tiers_masters_total.push(replicaset.current_master_name.clone());
             }
             continue;
         };
@@ -157,8 +197,27 @@ fn handle_truncate_table<'i>(
         if replicaset.tier == tier {
             tier_masters_count += 1;
         } else {
-            other_tiers_masters.push(master.name.clone());
+            other_tiers_masters_total.push(master.name.clone());
         }
+    }
+
+    let mut other_tiers_masters_batch = vec![];
+    if
+    // As first substep we do the vshard map_callrw for the actual
+    // implementation of TRUNCATE. And only after it was done we send the
+    // dummy proc_apply_schema_change RPC to other tiers' masters so that
+    // they bump their schema versions
+    last_step_info.truncate_map_callrw_ok
+        // get_next_batch expects a non-empty input
+        && !other_tiers_masters_total.is_empty()
+    {
+        let res = get_next_batch(&other_tiers_masters_total, last_step_info, batch_size);
+        other_tiers_masters_batch = match res {
+            Ok(v) => v,
+            Err(next_try) => {
+                return Ok(Some(SleepDueToBackoff::new(next_try, step_kind).into()));
+            }
+        };
     }
 
     let rpc = rpc::ddl_apply::Request {
@@ -172,12 +231,14 @@ fn handle_truncate_table<'i>(
     let success_cas = cas::Request::new(Op::DdlCommit, predicate, ADMIN_ID)?;
 
     Ok(Some(
-        ApplyTruncateTable {
+        Action {
+            schema_version,
             pending_schema_version,
             tier,
             rpc,
             tier_masters_count,
-            other_tiers_masters,
+            other_tiers_masters_total,
+            other_tiers_masters_batch,
             success_cas,
         }
         .into(),
@@ -258,49 +319,71 @@ pub fn collect_vshard_map_callrw_proc_apply_schema_change(
 /// space. This is a side-effect of using DdlPrepare/DdlCommit raft entries
 /// for the truncate operation.
 ///
+/// This function waits until responses to every request are received (or
+/// timeout is exceeded). `last_step_info` is updated with with info about
+/// results of every RPC, so that it can be used for the purposes of batching
+/// and back-off on following iterations.
+///
 /// Truncate operation can never result in a `DdlAbort` so if this function
 /// returns any error it is safe to attempt the exact same RPC later.
 pub async fn collect_proc_apply_schema_change(
-    targets: &[InstanceName],
+    last_step_info: &mut LastStepInfo,
+    targets: Vec<InstanceName>,
     rpc: &rpc::ddl_apply::Request,
     pool: &ConnectionPool,
     rpc_timeout: Duration,
-) -> Result<Result<Vec<()>, OnError>> {
-    let mut fs = vec![];
+) -> Result<(), OnError> {
+    let mut fs = FuturesUnordered::new();
     for instance_name in targets {
         tlog!(Info, "calling proc_apply_schema_change"; "instance_name" => %instance_name);
-        let resp = pool.call(
-            instance_name,
-            proc_name!(proc_apply_schema_change),
-            rpc,
-            rpc_timeout,
-        )?;
-        fs.push(async move {
-            match resp.await {
-                Ok(rpc::ddl_apply::Response::Ok) => {
-                    tlog!(Info, "applied schema change on instance";
-                        "instance_name" => %instance_name,
-                    );
-                    Ok(())
-                }
-                Ok(rpc::ddl_apply::Response::Abort { cause }) => {
-                    tlog!(Error, "failed to apply schema change on instance: {cause}";
-                        "instance_name" => %instance_name,
-                    );
-                    Err(OnError::Abort(cause))
-                }
-                Err(e) => {
-                    tlog!(Warning, "failed calling proc_apply_schema_change: {e}";
-                        "instance_name" => %instance_name
-                    );
-                    Err(OnError::Retry(e))
-                }
-            }
-        });
+        let resp = pool
+            .call(
+                &instance_name,
+                proc_name!(proc_apply_schema_change),
+                rpc,
+                rpc_timeout,
+            )
+            .map_err(OnError::Retry)?;
+        fs.push(async move { (instance_name, resp.await) });
     }
 
-    let res = try_join_all(fs).await;
-    Ok(res)
+    let mut abort_cause = None;
+    let mut first_error = None;
+
+    while let Some((instance_name, res)) = fs.next().await {
+        match res {
+            Ok(rpc::ddl_apply::Response::Ok) => {
+                tlog!(Info, "applied schema change on instance"; "instance_name" => %instance_name);
+                last_step_info.on_ok_instance(instance_name);
+            }
+            Ok(rpc::ddl_apply::Response::Abort { cause }) => {
+                tlog!(Error, "failed to apply schema change on instance: {cause}"; "instance_name" => %instance_name);
+                // For now we only keep the error from the first failed
+                // instance but maybe we should save them from all of them
+                if abort_cause.is_none() {
+                    abort_cause = Some(cause);
+                }
+            }
+            Err(e) => {
+                let info = last_step_info.on_err_instance(&instance_name);
+                let streak = info.streak;
+                tlog!(Warning, "failed calling proc_apply_schema_change (fail streak: {streak}): {e}"; "instance_name" => %instance_name);
+                if first_error.is_none() {
+                    first_error = Some(e);
+                }
+            }
+        }
+    }
+
+    if let Some(cause) = abort_cause {
+        return Err(OnError::Abort(cause));
+    }
+
+    if let Some(error) = first_error {
+        return Err(OnError::Retry(error));
+    }
+
+    Ok(())
 }
 
 ////////////////////////////////////////////////////////////////////////////////